+++ Common/src/main/java/fuzs/resourcepackoverrides/mixin/client/PackSelectionScreenMixin.java
@@ -1,7 +1,8 @@
 package fuzs.resourcepackoverrides.mixin.client;
 
-import fuzs.resourcepackoverrides.client.gui.screens.packs.PackAwareSelectionEntry;
+import fuzs.resourcepackoverrides.client.data.PackSelectionOverride;
 import fuzs.resourcepackoverrides.client.data.ResourceOverridesManager;
+import fuzs.resourcepackoverrides.client.gui.screens.packs.PackAwareSelectionEntry;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.packs.PackSelectionModel;
 import net.minecraft.client.gui.screens.packs.PackSelectionScreen;
@@ -22,6 +23,12 @@ abstract class PackSelectionScreenMixin extends Screen {
     @ModifyVariable(method = "updateList", at = @At("HEAD"))
     private Stream<PackSelectionModel.Entry> updateList(Stream<PackSelectionModel.Entry> models) {
         // This also runs on the data packs screen, but that's fine since the entries are not wrapped there.
-        return models.filter(entry -> !(entry instanceof PackAwareSelectionEntry contextEntry) || !ResourceOverridesManager.getOverride(contextEntry.getPackId()).hidden());
+        return models.filter(entry -> {
+            if (!(entry instanceof PackAwareSelectionEntry contextEntry)) {
+                return true;
+            }
+            PackSelectionOverride override = ResourceOverridesManager.getOverride(contextEntry.getPackId());
+            return override.hidden() == null || !override.hidden();
+        });
     }
 }