+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v3.0.1-1.18.2] - 2023-03-14
+### Added
+- Additional resource pack attributes for overriding: `title`, `description`, `default_position`
+### Changed
+- Some attributes are no longer ignored when set to `false`
+### Fixed
+- Fixed crash on game launch on Forge due to invalid Mixin injector
+
 ## [v3.0.0-1.18.2] - 2023-02-28
 - Ported from 1.19 version